requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
tensorflow
Boruta
keras-tcn